safari-600.7-branch/LayoutTests/http/tests/security/dataURL/xss-DENIED-to-data-url-in-foreign-domain-subframe-location-change-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

Opener frame


PASS: Cross frame access to a data: URL embed in a frame on a foreign domain denied!


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS: Cross frame access from a frame on a foreign domain denied!

Inner iframe. This iframe (which is data: URL and whose parent is on a foreign domain) is the frame that the main frame is trying to access. It should not have access to it.
